Added increased timeout on integration test for account view (#554)
no refs

When the account view changes were merged we started seeing flaky tests
due to timeouts. Looking at the tests, there's not a whole lot that can be
optimized to try and reduce the likelihood of timeouts. For this set of tests,
we have increased the timeout to 10 seconds, which is a similar approach to
we have taken for the integration tests for the feed service