Fixed performance of viewByApIdRemote (#575)
ref https://linear.app/ghost/issue/PROD-677

The `ap_id` column is not indexed due to the size of it, instead we have
the `ap_id_hash` column which contains the SHA256 of the `ap_id` and
is indexed. We should never be adding WHERE conditions on the `ap_id`
columns in the codebase, it should always be via the indexed hash
instead.