Fixed flaky moderation service test
ref ref https://linear.app/ghost/issue/AP-1082

Added fix for a flaky moderation service test - The flakiness was being caused
due to using the wrong field when returning the list of filtered out users
when the author of a post has blocked the reposter. The previously returned
field would just happen to match the expected value most of the time, which
caused the test to fail intermittently. This fix ensures that the correct field
is used to filter out users and prevents the test from failing intermittently.