Used `Account` type instead of `AccountEntity`
We're going to update the AccountEntity to have infra specific data on it,
causing it to diverge from the Account interface. These tests actually want to
be working with the Account interface and don't care about the Entity impl, and
we should also depend on the smallest or most specific type when possible.