Supported raising events at the AccountEntity level
ref https://linear.app/ghost/issue/PROD-669

We are moving toward having the domain implementation handle the creation of
events, rather than the repository. This requires that we have a way for an
entity to create events that will later be dispatched by the infra layer.

We're able to do this only on the AccountEntity, without polluting the Account
interface - keeping it focused purely on the domain.