Added `block(account: Account)` capability to Account
ref https://linear.app/ghost/issue/PROD-669

This just adds the AccountBlockedEvent to be dispatched. This event will be
handled and have the block persisted to the database. This allows us to avoid
maintaining a list of blocks on the AccountEntity - but we do lose the ability
to be able to error if the account is already blocked.