Stored current Account on AppContext via middleware
We have a lot of boilerplate code to load the Account for the current request,
which generally calls either `getBySite` or `getDefaultAccountForSite` methods.
We want to deprecate these so that our controller layer does not need to call
these methods directly and also so that when we move to multiple accounts per
site, we are able to handle loading of the correct Account in a single place.