Switched AccountBlockedEvent to use account ids
ref https://linear.app/ghost/issue/PROD-669

We've been thinking about moving events to using ids for a little while now
because it reduces their size and makes them more appropriate to send over a
queue like PubSub. Now that the Account entity is immutable it makes even more
sense to use the id rather than the entity instance, because the entity could
have been updated, and we'd have a reference to the old one.