Asserted that an Account is updated when saving
ref https://github.com/TryGhost/ActivityPub/pull/583

This was a suggestion by CodeRabbit in the linked PR. We can never get an
Account entity that doesn't have a corresponding id in the DB - so this should
never happen, and if it does something is seriously wrong. Throwing an error
here is in line with out fail fast philosophy and well bubble up the stack and
result in a 5xx - alerting us to the unexpected behaviour.