Increased test timeout to 10 seconds (#600)
no ref

We are noticing tests failing on a regular basis on CI due to regularly hitting
the preconfigured timeout limit - We think this is down to resoruce constraints
on the CI + other potentia environmental factors (network requests flowing
between services in the e2e tests). This PR ups the timeout to 10 seconds to
see if it help alleviates the regularity of test timeouts.