Refactored repost controller logic into service
ref https://linear.app/ghost/issue/PROD-669

By moving this logic all into the service we have a single place to add
restrictions based on whether or not an account is blocked. We've also moved to
the new Result type - which forces us to handle errors explicitly in the
controller. The structure of the Activity we send has changed slightly, with
just the object ID instead of the full object now - which has meant our tests
had to be updated.