Added `followedByMe` to AccountFollow DTO
The more aligned all of our DTOs are, the better, because we can share code and
queries to return the same objects. This is adding the `followedByMe` field so
that we can later remove the more ambiguous `isFollowing` field. We're not
doing a wholesale replacement so that the client has time to update.