Added APIs for reading blocked accounts & domains
ref https://linear.app/ghost/issue/PROD-1636

This supports our new UI of listing all the blocked accounts and domains and
having the ability to unblock them if desired.

The BlocksView follows our existing pattern of skipping the services &
repositories for queries. We also want to maintain consistency with other API
endpoints, so we now have a shared MinimalAccountDTO which is used for follows
as well as blocked accounts.