Removed call to `getActor()` on Create activities
ref https://linear.app/ghost/issue/PROD-1587

This would sometimes throw, and would cause us to respond with a 500.
The `sender` was only used for the legacy key_value inbox, which is only
used in tests at the moment. We've maintained the functionality of the
inbox, but are relying on the `post.author` instead of `create.getActor()`
to determine whether a follow relation exists.

If there is an error fetching the underlying Actor - this will be caught in the
`getByApId` method of the `postService` and handled in switch statement

We've also renamed the params to `checkIfAccountIsFollowing` because
they tripped me up when making this change.

Co-authored-by: Fabien O'Carroll <fabien@allou.is>