Added post stats to notifications API (#668)
ref https://linear.app/ghost/issue/PROD-1668

We want to update the notifications UI to display reply notifications as a
post, similar to how the appear in the feed, with the ability to interact with
the post directly from notifications.

The joins to the `likes` and `reposts` table have been done in the order of
matching `account_id` followed by the `post_id` to take advantage of the
composite index we have on the columns in this order.