Reduced support for the ActivityPub Outbox
This is still reliant on the key_value table, which we want to remove support
for. As a first step we are reducing the support for the outbox, which isn't
used by our application, and isn't essential for federation. We still support
the endpoint, but it always returns an empty collection.

Our e2e cucumber tests had a lot of reliance on the outbox, so the tests and
step defs have been updated to preserve features and to keep the suite passing.