Fixed and added regression tests for 5xx errors
ref https://github.com/fedify-dev/fedify/issues/232

This is a large source of 500's in production for us, and they're out
of our control. Until this is fixed in Fedify, or we have a better way of
handling the errors we have to explicitly catch the errors in the global
Hono error handler. The regression tests helped to replicate the
problem, and will help with validating the fix in Fedify in future.