Fixed `authoredByMe` property when reading local posts
ref https://linear.app/ghost/issue/PROD-1638

We were incorrectly passing `accountId` instead of `defaultAccountId` to the
`mapPostToDTO` method, which resulted in always having true for `authoredByMe`.

This meant that we showed the "delete" button on posts which aren't ours.

I've renamed `defaultAccountId` to `contextAccountId` so it's clearer what it
is, added JSDoc to the `mapPostToDTO` method to explain the parameter and fixed
which param we pass. I've also fixed the test which should have caught this to
expect the correct data.