Updated Actor dispatcher to use identifier
ref https://linear.app/ghost/issue/PROD-1775

Our actor dispatcher now correctly uses identifier, so that 'index' is the
identifier. We then have a username/handle mapper which is currently hardcoded
to 'index'. This allows us to remove a lot of the hardcoding elsewhere in the
application, so that we can correctly pass `username`.

Now we're in a good position to move toward multiple accounts, as we will just
need to update the `mapHandle` and `actorDispatcher` to correctly fetch from
the DB and the `sendActivity` calls will all work as expected.