Handled "duplicate" post.published webhooks
closes https://linear.app/ghost/issue/PROD-1905

When we recieve a second webhook for the same post, we attempt to save another
post with the same uuid - which throws a duplicate key error. We should really
be doing some kind of update on the post, but we don't currently support that
and it's out of scope of fixing the error.

Instead we check for the existence of the post first, and return an error.