Used awilix to dependency inject services
ref https://linear.app/ghost/issue/PROD-1662
    
Up until now we have been manually injecting our dependencies, which is
cumbersome. We've wanted to move towards a systematic approach for a while but
haven't gotten around to it. With self hosting support on the horizon it is
becoming more important that we have a way to configure which dependencies are
loaded and a DI container system makes that fairly simple.
    
Right now this is just as verbose as it previously was, but it moves us in the
right direction. Eventually we will move the container configuration into a
separate file, as well as provide helper functions for wiring up endpoints &
fedify handlers to service implementations.