Implemented `getAncestors` for ReplyChainView
ref https://linear.app/ghost/issue/PROD-1935

This is our first foray into using recursion in MySQL. What we're doing is
building up a small in-memory table of the post ids we're interested in, and
then selecting from the `posts` table based on a join to this in-memory table.

The in-memory table is created by recursively joining `posts` onto itself via
the `in_reply_to` column and the recursion is limited by a clause on `depth`

We've also abstracted out the selection of a post row into a separate method,
as we're going to need to reuse then when reading children and their replies.