Implemented `getChildren` for `ReplyChainView`
ref https://linear.app/ghost/issue/PROD-1935

This is a little tricker than the previous one, we use the same approach of
constructing an in-memory table which is then joined to posts to fetch them.

We construct the table by first selecting the top level children using LIMIT to
restrict the count, and then the recursion happens in a similar way to the
ancestors, except we are climbing down the chain, and only if reply_count is 1.

The depth is limited in the same way as the ancestor query.

The in-memory table uses the ROW_NUMBER window function over created_at of the
top level comments, this is then shared with all of the replies in its chain.

This allows us to order the results first by the parent created at, and then by
the depth, so that we get an order like:

 - 1
 - 1.1
 - 1.1.1
 - 2
 - 2.1
 - 3
 - 4
 - 4.1

Which makes reconstructing theses results into our desired data structure
possible in a single loop.