Moved AccountUpdatedEvent into Entity
ref https://linear.app/ghost/issue/PROD-1986

We don't want to emit an AccountUpdatedEvent for every time an entity is saved,
we only want to do it when we have actually updated the data. This may also
have knock-on effects where an account was updated, but in the context of
another accounts request, and the event handler was attempting to call
`sendActivity` but the keys used were from the wrong context, so it fails.