Filtered on `ap_id_hash` instead of `ap_id` (#805)
ref https://linear.app/ghost/issue/PROD-1969

The `ap_id` column is not indexed, which means we need to do a full table scan
when filtering on it, which kills performance. Instead we have the `ap_id_hash`
column for this filter, which is indexed.

Original query in production takes ~1s, this one takes ~0.2ms - 5000x faster