Replaced `next` with `hasMore` for ancestors & reply chains
We're not using a new APi to fetch more ancestors or reply chain depth, instead
we're using the existing API and just starting at a different point, higher or
lower in the chain. Because of this the concept of `next` doesn't make sense -
it's not a parameter we can pass. Instead we're going to use a `hasMore`
property which indicates to the client if it's worth fetching more data.

We didn't end up needing separate API's to fetch more ancestors/reply chains,
as we can just use the existing API