Removed the host-scoped KVStore
no-issue

We are no longer writing to a scoped KVStore! We only write to it in a global
context to e.g. store activities or the cache for fedify, or cached JWKS!

This means we can remove it from both the Hono and the Fedify context, and then
clean up a bunch of unused code too.