Added initial populate-explore-json job
ref https://linear.app/ghost/issue/PROD-2011

We want to be able to generate a static JSON file for populating the
accounts for the explore page. The idea here is to run this job in the
background and update it periodically to keep counts and order in sync.

This is just to get the thing wired up, and then we can start
implementing it afterwards