Fixed cucumber tests for notification counts
no-issue

The previous tests were not waiting for the notification to be present
before checking for the counts. I've updated the tests to be a little
more focused on what we're testing, and pushed some complexity into the
step definitions, rather than adding extra steps which verify the
notification has been received.

Note that we pass a custom matcher when finding like notifications
because matching on the post.id isn't unique enough, as it refers to
the post which was liked.