Refactored DI registrations to separate module (#855)
no-issue

This doesn't contain any functional changes, it's pretty much copy &
paste from app.ts into a function in configuration/registrations.ts

We've wrapped the registrations in a function so we can (in-theory) test
it in the future.

We've had to rename a few globals in app.ts to stop the build process
from clobbering parameter names.