Removed transactions from SiteService
no-issue

First, we don't want to be dealing with the database at the service
level, let alone transactions. Secondly - they are interacting with
transactions at the repository layer in a confusing way.

Because we removed the transactions, we need to explicitly handle
partial states because it's now possible for only some of the operations
to have completed.