Moved event spy until after test setup
Adding the spy before setting up the state for a test means that if the
state setup is refactored, e.g. creation of accounts, it can alter the
state of the spy in a way that is not relavant to the test.

This changes makes the tests less brittle to change and prepares for the
AccountCreatedEvent to be emitted when creating internal accounts.