Added GhostExploreServic
The Ghost Explore service handles having Ghost accounts follow
the explore mastodon account, so we can easily find and promote
their content.

This also updates the repository to emit the events outside of the
transaction, so that the handlers have the committed db rows available.

We also move the middleware to register the fedify context as we will
need the context when handling the site data endpoint, where the
internal account for a site is created.