Added check for Account before issuing follow
If Mastodon receives a follow from an account and the account is not
loading, it will blacklist the ip which made the request for 5 minutes.

This cause our script to have a lot of failures, so we're checking
ourselves first if the account is reachable, before issuing the follow.

I've also tried to clean up the logging to make it easier to see what's
going on.