Added check for Account before issuing follow (#953)
When Mastodon receives a follow from an account it hasn't seen before, 
it attempts to go and get their public key. If the request for the
public key fails for some reason — for example, SSL is misconfigured or
the domain doesn't exist anymore — Mastodon will banish that IP for five
minutes.

This cause our script to have a lot of failures, so we're checking
ourselves first if the account is reachable, before issuing the follow.

I've also tried to clean up the logging to make it easier to see what's
going on.