Ensured posts from announced `Create` activities are reposted
ref https://linear.app/ghost/issue/PROD-2227

Announced `Create` activities usually come from `Group` actors announcing
`Create` activities from members of the group. If you followed the group and
not the author account, you would not see the announced posts from the
group in your feed (because you are not following the author account and this
is checked when adding to the feed). These are now correctly reposted by the
group account so they can appear in the feeds of followers of the group

This changeset also includes a dev script for spinning up a local WordPress
instance for testing interop