Added fedify delivery error analysis to prevent redudant retries
ref https://linear.app/ghost/issue/PROD-2210

Added error analysis when processing a message from the queue. If the error is
a Fedify delivery error (classified based on the format of the message), we
analyze the status code in the message and decide on a retry / report strategy
based on this status code. In the future we can also analyze other types of
error to decide on an applicable retry / report strategy for the given error