Added constraints to title, summary & excerpt
ref https://linear.app/ghost/issue/PROD-2272

We want to be able to model constraints in the type system, so that we
can be sure that they are handled everywhere. If we're to continue to
use the string type everywhere, we can't be sure that we don't forget
validation.

Rather than using fully fledged Value Objects for a simple constraint
like this, I've opted for "branded" types - which are a little more
lightweight. The idea is that you can only get this type from the parse
method, but we can use it as a string, without needing to unwrap some
kind of PostSummary class instance.