Fixed potential bug in setting anomaly retryable state to false
Shouldn't be an issue, since it's only possible to set this boolean when the anomaly is a "maybe retryable" one. In those cases, the default is false, so passing false will get the default. But it's still better to actually use the user-input instead of falling back to the `getRetryable` method.

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>