Improved tracing for HTTP & PubSub
The spanWrapper is being used with anonymous functions which means
that the names are all 'anonymous'. The controllers have been updated to
have explicit names now. Also the Sentry middleware has been simplified
and moved into the http integration, which was already enabled - so this
deduplicates spans for http requests, as well as improves the naming.

Finally the PubSub has spans explicitly emitted now, and there's been some
wrangling to get the internal fedify spans linked correctly!