Handled FetchErrors in Inbox handlers
ref https://linear.app/ghost/issue/PROD-2285

Errors when processing Inbox activities are surfaced via the inbox
error handler, as well as the message queue error handler when
running with a queue. This means that for each error in inbox handlers
we were reporting to Sentry twice. Now we are only going via the
MessageQueue and we've added extra handling for FetchError
which is thrown by methods like `getObject` and `getActor`