Simplified Activity type checking for filtering
ref https://linear.app/ghost/issue/PROD-2324

We only send standard Activity Streams activity types, so we are safe to
just match on the string instead of referring to the whole context url
stored on activityType. This also updates the types to reflect this, but
we have ther are optional as we're not certain they are always present.