Tested activity delivery to internal accounts
ref https://linear.app/ghost/issue/PROD-2324

We are about to completely change how we delivery activities to internal
accounts, rather than sending them over the network we're doing to rely
on the domain layer to implement all of the side effects. In order to
have confidence with this change I wanted to test the flows explicitly.

Rather than a new internal accounts feature file, I want to make sure
that feature files are broken down by feature, and delivery to internal
accounts is just a part of each feature. Though the internal account
steps are all together because they are all of the same newer style, and
they are all quite similar and can eventually share a lot more code.