Fixed bugs when delivering activities to inboxes
ref https://github.com/fedify-dev/fedify/pull/335
ref https://github.com/fedify-dev/fedify/pull/324

This bump to fedify includes two fixes the the doubleKnock
implementation. One which is an attempt to remedy the 'unusable' errors
when cloning the request objects, and the other which handles relative
paths in the Location header when recieving a redirect.

Both of these errors cause significant noise in our production
environment and cleaning them up should allow us to spot real errors a
lot easier in future.