Ensured that ActivityPub endpoints can be cached
ref https://linear.app/ghost/issue/PROD-2331

Surrogates like Fastly will not cache requests if they are marked as
private, regardless of the Surrogate-Control header.

Marking this as public ensures that the content can be cached by
surrogates, and setting a max-age of 0 ensures that browsers don't
cache it, giving us full control of purging at the surrogate level.