Added handling for a site being disabled / enabled (#1132)
ref https://linear.app/ghost/issue/PROD-2351

When a site is disabled (via `DELETE` request to `/v1/sites`) the `site` record
will be removed from the database. This will cascade and remove the `user`
associated with the site, the feed entries associated with the user and the
notifications associated with the user. This will leave the `account` in-tact.
If the site is later re-enabled, a new site & user record is generated. The
user record will be linked to the previously left over account. If there was
a previously left over account, it will retain its keypair in case other servers
have cached this