Added support for RedisKvStore for Fedify
Fedify uses the KvStore a lot and often with an expiry. Currently the
MySQL backed KVStore can get overwhelmed with a lot of inserts, which
can cause connections to be dropped which affects uptime for the core
application. Redis also supports TTL out of the box, so this eradicates
the need for a cleanup job running on the key_value table.