Fixed reading preloaded contexts from KvStore
ref https://github.com/fedify-dev/fedify/pull/352

The `kvCache` wrapper for document loaders was reading straight from the
cache, even for preloaded contexts, which mean a large number of DB
queries being made for data that is in-memory in the containers.

This patch is a fix for this behaviour which will bypass the cache
completely when loading preloaded contexts documents.