Upgraded biome to v2.1.3 (#1162)
no ref

Upgraded biome to the latest version. As this was a major release, some of the
rules / formatters have been updated / changed which has meant the code has
had to change also. There are no logic changes, the main changes are related to
ordering imports, removing unused variables or renaming unusued but ordered
variables