Included pool information on Connection lost error
ref https://linear.app/ghost/issue/PROD-2530

We're pretty sure that the connection pool idle time was the cause of
the connection lost errors, but we want to add some extra information
to the errors, if this does crop up again, we'll have more information
about the current state of the system.

This code is not long term and will be removed once we're confident
that the connection errors are solved.