Added bluesky handle confirmation endpoint
no ref

We found situations where the computed Bluesky handle did not match what Bridgy
had actually used for the account - This is because Bridgy has some
conditional logic based on the domain setup. We do not want to emulate this
logic so we have instead changed our approach to require handle confirmation.
The client will request handle confirmation in which we will look up the
Bluesky account that should have been created for the account and use the actual
handle that was assigned