Fixed `slack-notify` job not running on workflow failure (again)
ref https://github.com/TryGhost/ActivityPub/pull/1324

Apparently the previous fix did not work, but it's not totally clear why. There
seems to be an issue using `always()` and `failure()` together.

According to the official GitHub docs, `failure()` should return true
`when any ancestor job fails,` but in practice (as documented in
https://github.com/orgs/community/discussions/80788),
it always returns false even when dependent jobs fail.

Have changed condition to explicitly check on the results of the previous jobs