Fixed site creation failure due to duplicate `ghost_uuid` entry
no ref

It is possible for a multiple sites to have the same `ghost_uuid` - i.e they
are the same site, but the domain has changed. Currently if a site is registered,
and its domain is changed, when it tries to register with the service, an error
is thrown due to the unique constraint on `sites.ghost_uuid`. This change fixes
this issue by first nullifying the `ghost_uuid` on the existing site before
creating the new site